pyproject.toml
[tool.poetry]
name = "coder-gym"
version = "0.1.0"
description = ""
authors = ["jshuckbot <gorka_ivan@mail.ru>"]
readme = ["README.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
pillow = "^10.1.0"
django-braces = "^1.15.0"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.0"
autoflake = "^2.2.1"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
